// src/types/api.ts
export interface DiscordActivity {
  details: string;
  state: string;
  startTimestamp: number;
  largeImageKey: string;
  largeImageText: string;
  smallImageKey: string;
  smallImageText: string;
  instance: boolean;
}